challenge_5/requirements.txt
azure-functions
azure.cognitiveservices.language.textanalytics
msrest.authentication
